// src/pacientes/dto/create-paciente.dto.ts
export class CreatePacienteDto {
    codigo_postal: string;
    Delegacion: string;
    direccion: string;
    es_usuario_principal: string;
    Estado: string;
    fecha_nacimiento: Date;
    genero: string;
    MunicipioCiudad: string;
    nro_documento: string;
    Pais: string;
    telefono_fijo: string;
    telefono_movil: string;
    tipo_documento: string;
}